opusrtp: Consolidate sniff and extract code
Add -o/--output to specify output file for either sniff or extract
and allow - for stdout.  Allow sniff/extract console output with no
output file (do not default to writing a file with a fixed name).
Use -r/--rate instead of -s/--samplerate for consistency with opusdec.
Clearly indicate in --help which options require an argument.